requirements.txt
altair==5.0.1
google_api_python_client==2.90.0
matplotlib==3.7.1
pandas==2.0.2
pgbm==2.1.1
protobuf==4.23.3
requests==2.31.0
scipy==1.10.1
seaborn==0.12.2
streamlit
selenium==4.10.0
torch==2.0.1
gurobipy==10.0.2
numpy==1.25.0
PyYAML~=6.0
supabase~=1.0.3
python-dotenv~=1.0.0